+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 
 - Add a workflow demo (caravan.examples.demo)
 - Changed Workflow api for defaults
+- Support workflow default timeouts as type int
 
 
 0.2.1 (2015-12-07)